Differential Revision: D14078519
Differential Version: 73345052

+++ caffe2/contrib/aten/aten_op.cc
@@ -14,7 +14,7 @@ OPERATOR_SCHEMA(ATen);
 namespace math {
 template <>
 void Set<at::Half, CPUContext>(
-    const size_t /*N*/,
+    const int /*N*/,
     const at::Half h,
     at::Half* v,
     CPUContext* c) {
